Incorporating channelSignalReconstructor
Moving the impulsivity calculating to the channelSignalReconstructor module and incorporating this module in the stationLDAVariables run function
